Move `auto_kill_aiosqlite` fixture to `conftest`
