+++ tests/test_appdb.py
@@ -1,8 +1,6 @@
 import asyncio
-import contextlib
 from datetime import UTC, datetime, timedelta
 import pathlib
-import threading
 import time
 
 import aiosqlite
@@ -46,26 +44,6 @@ from zigpy.zcl import ClusterType, UnsupportedAttribute
 from zigpy.zcl.clusters.general import Basic, Identify, OnOff, Ota
 from zigpy.zcl.foundation import Status as ZCLStatus, ZCLAttributeDef
 from zigpy.zdo import types as zdo_t
-
-
-@pytest.fixture(autouse=True)
-def auto_kill_aiosqlite():
-    """Aiosqlite's background thread does not let pytest exit when a failure occurs"""
-    yield
-
-    for thread in threading.enumerate():
-        if not isinstance(thread, aiosqlite.core.Connection):
-            continue
-
-        try:
-            conn = thread._conn
-        except ValueError:
-            pass
-        else:
-            with contextlib.suppress(zigpy.appdb.sqlite3.ProgrammingError):
-                conn.close()
-
-        thread._running = False
 
 
 async def make_app_with_db(database_file):
